@@ -7,8 +7,8 @@
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "dompurify": "^2.3.6",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-router": "^6.2.1",
     "react-scripts": "^2.1.3",
     "react-transition-group": "^4.4.2",
