@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useTransition } from 'react';
 import { CSSTransition, SwitchTransition } from 'react-transition-group';
 import contactBackend from '../Utilities/contactBackend';
 import createMarkup from '../Utilities/createMarkup';
@@ -7,12 +7,17 @@ import Resume from './Resume';
 
 const Frame = (props) => {
     const [content, setContent] = useState(''); //state to hold markup
+    const [isLoading, startTransition] = useTransition();
     const [animSwitchF, setAnimSwitchF] = useState(false); //animation switch for CSSTransition
 
     /* the useEffect hook on the viewF state here sets the content state with fetched html and data
     on the components unmount the css transition is set to false so it can be cleanly toggled for the new state*/
     useEffect(() => {
-        contactBackend(props.viewF, setContent);
+        setContent({htmlPack:"<span>loading</span>", type:"loading"})
+
+        startTransition(() => {
+            contactBackend(props.viewF, setContent);
+        })
 
         return () => {
             setAnimSwitchF(false);
@@ -39,7 +44,7 @@ const Frame = (props) => {
 
     return (
         <SwitchTransition>
-            <CSSTransition key={props.viewF} in={animSwitchF} timeout={600} classNames="my-frame" unmountOnExit>
+            <CSSTransition key={props.viewF} in={content} timeout={600} classNames="my-frame" unmountOnExit>
                 <div id="mainView" className="m-2" >
                     <section id="content" dangerouslySetInnerHTML={createMarkup(content.htmlPack)}></section>
                     {secondaryComponent(content.type)}
